chireads:  - refactored for plugin beta  - separted parseNovels  - separted getCheerio  - loadedCheerio => $ for shorter lines & simpler test on web
